requirements.txt
# Flask and core web framework
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==22.0.0

# Database and caching
mysql-connector-python==8.3.0
redis==5.2.0

# WeChat integration
wechatpy==1.9.12

# API and HTTP requests
requests==2.32.3

# JWT authentication
PyJWT==2.10.1

# AI and machine learning
openai==1.57.0

# Data processing and manipulation
pandas==2.2.3

# Chinese date parsing
cn2date==0.1.2

# Model Context Protocol
mcp==1.3.0

# Development and testing (optional)
python-dotenv==1.0.1
